8:32 AM fixed _REGISTER_TRACKERS_DIST; 11:18 AM fixed quite a few warnings; 2:24 PM updated all SSMs for unsigned int; 4:07 PM completed fixing warnings in the main modules; 4:13 PM completed initial fixing of warnings in parameters.h too; 4:41 PM updated MIL and TLD for windows; 5:07 PM updated MIL to avoid some of the naming conflicts; 5:32 PM finally got TLD cmake file working in windows; 5:41 PM finally got MIL compiling; 5:46 PM got TLD compiling too; 10:48 PM fixed a few warnings in mtf.h;

+++ AM/src/ImageBase.cc
@@ -33,8 +33,8 @@ uchar_input(UCHAR_INPUT){
 		if(img_params->resx <= 0 || img_params->resy <= 0) {
 			throw std::invalid_argument("ImageBase::Invalid sampling resolution provided");
 		}
-		resx = img_params->resx;
-		resy = img_params->resy;
+		resx = static_cast<unsigned int>(img_params->resx);
+		resy = static_cast<unsigned int>(img_params->resy);
 		n_pix = resx*resy;
 		grad_eps = img_params->grad_eps;
 		hess_eps = img_params->hess_eps;
